moved SchemaFactory
