+++ v12x/src/test/java/com/foursoft/harness/vec/v12x/BasicLoadingTest.java
@@ -10,10 +10,10 @@
  * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  * copies of the Software, and to permit persons to whom the Software is
  * furnished to do so, subject to the following conditions:
- * 
+ *
  * The above copyright notice and this permission notice shall be included in
  * all copies or substantial portions of the Software.
- * 
+ *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -25,6 +25,7 @@
  */
 package com.foursoft.harness.vec.v12x;
 
+import com.foursoft.harness.vec.v12x.validation.SchemaFactory;
 import com.foursoft.jaxb.navext.runtime.ExtendedUnmarshaller;
 import com.foursoft.jaxb.navext.runtime.JaxbModel;
 import com.foursoft.jaxb.navext.runtime.io.utils.ValidationEventLogger;
